[image-picker] Extracted `ImageOrVideoPreview` component into separate file
